about to try to sed my way out of trouble
I mixed up variable naming convention for buy/sell functionality.
Some is in the perspective of the player buying/selling, some is
in reference to what the current Hero wants to do.  Quite a mess.
Before trying to sed all my variables to a logical state, I'm
saving this current state of the project.